HSEARCH-3735 Move failure context building to API
We need to if we want to allow automatic indexing synchronization
strategies to have access to the FailureHandler.
