+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/orchestration/impl/LuceneEnsureIndexExistsWriteWorkSet.java
@@ -8,7 +8,7 @@ package org.hibernate.search.backend.lucene.orchestration.impl;
 
 import java.util.concurrent.CompletableFuture;
 
-import org.hibernate.search.engine.reporting.spi.IndexFailureContextImpl;
+import org.hibernate.search.engine.reporting.IndexFailureContext;
 
 /**
  * A special workset that won't trigger the creation of the index writer.
@@ -32,7 +32,7 @@ class LuceneEnsureIndexExistsWriteWorkSet implements LuceneWriteWorkSet {
 		catch (RuntimeException e) {
 			markAsFailed( e );
 			// FIXME HSEARCH-3735 This is temporary and should be removed when all failures are reported to the mapper directly
-			IndexFailureContextImpl.Builder failureContextBuilder = new IndexFailureContextImpl.Builder();
+			IndexFailureContext.Builder failureContextBuilder = IndexFailureContext.builder();
 			failureContextBuilder.throwable( e );
 			failureContextBuilder.failingOperation( "Index initialization" );
 			processor.getFailureHandler().handle( failureContextBuilder.build() );
