+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/orchestration/impl/LuceneSingleWriteWorkSet.java
@@ -12,7 +12,6 @@ import org.hibernate.search.backend.lucene.work.impl.LuceneWriteWork;
 import org.hibernate.search.engine.backend.work.execution.DocumentCommitStrategy;
 import org.hibernate.search.engine.backend.work.execution.DocumentRefreshStrategy;
 import org.hibernate.search.engine.reporting.IndexFailureContext;
-import org.hibernate.search.engine.reporting.spi.IndexFailureContextImpl;
 
 class LuceneSingleWriteWorkSet<T> implements LuceneWriteWorkSet {
 	private final LuceneWriteWork<T> work;
@@ -39,7 +38,7 @@ class LuceneSingleWriteWorkSet<T> implements LuceneWriteWorkSet {
 		catch (RuntimeException e) {
 			markAsFailed( e );
 			// FIXME HSEARCH-3735 This is temporary and should be removed when all failures are reported to the mapper directly
-			IndexFailureContextImpl.Builder failureContextBuilder = new IndexFailureContextImpl.Builder();
+			IndexFailureContext.Builder failureContextBuilder = IndexFailureContext.builder();
 			failureContextBuilder.throwable( e );
 			failureContextBuilder.failingOperation( work.getInfo() );
 			IndexFailureContext failureContext = failureContextBuilder.build();
