+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/massindexing/impl/FailureHandledRunnable.java
@@ -8,8 +8,8 @@ package org.hibernate.search.mapper.orm.massindexing.impl;
 
 import java.lang.invoke.MethodHandles;
 
+import org.hibernate.search.engine.reporting.FailureContext;
 import org.hibernate.search.engine.reporting.FailureHandler;
-import org.hibernate.search.engine.reporting.spi.FailureContextImpl;
 import org.hibernate.search.mapper.orm.logging.impl.Log;
 import org.hibernate.search.util.common.logging.impl.LoggerFactory;
 
@@ -41,7 +41,7 @@ abstract class FailureHandledRunnable implements Runnable {
 			}
 
 			// being this an async thread we want to make sure everything is somehow reported
-			FailureContextImpl.Builder contextBuilder = new FailureContextImpl.Builder();
+			FailureContext.Builder contextBuilder = FailureContext.builder();
 			contextBuilder.throwable( re );
 			contextBuilder.failingOperation( log.massIndexerOperation() );
 			failureHandler.handle( contextBuilder.build() );
